Module 16/waste_project/requirements.txt
scikit-learn
tensorflow
keras
pillow
numpy